Server/requirements.txt
fastapi
uvicorn
yfinance
requests
beautifulsoup4
python-dotenv
pydantic
email-validator
openai
pandas
